separate indicator view from avatar on LHN header
